requirements.txt
flask
torch
torchvision
numpy
opencv-python
Pillow
scikit-learn
tqdm
matplotlib
librosa
soundfile
tensorflow
